Feature/issue 2283 s3 porting part1 - 25 (#2896)
Porting ClickHouse #37964，#43253，#43824，#41979，#36647，#42376，#43454，#38196，#34539 and #43726，#43667，#34566，#37259，#38542，#41652，#42628，#45442，#42767，#41437，#42800，#44289，#42190, #36312. 25th pr of #2283

+++ src/Common/examples/parallel_aggregation.cpp
@@ -501,7 +501,7 @@ int main(int argc, char ** argv)
 
         watch.restart();
 
-        for (size_t i = 0; i < MapTwoLevel::NUM_BUCKETS; ++i)
+        for (unsigned i = 0; i < MapTwoLevel::NUM_BUCKETS; ++i)
             pool.scheduleOrThrowOnError([&] { merge2(maps.data(), num_threads, i); });
 
         pool.wait();
@@ -554,7 +554,7 @@ int main(int argc, char ** argv)
 
         watch.restart();
 
-        for (size_t i = 0; i < MapTwoLevel::NUM_BUCKETS; ++i)
+        for (unsigned i = 0; i < MapTwoLevel::NUM_BUCKETS; ++i)
             pool.scheduleOrThrowOnError([&] { merge2(maps.data(), num_threads, i); });
 
         pool.wait();
